# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Gestión de Tareas",
    "version": "1.0.0",
    "description": "API para la gestión de usuarios y tareas, incluyendo funcionalidades de autenticación y procesamiento de tareas con IA."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Servidor de desarrollo local"
    },
    {
      "url": "http://0000243.xyz:8080",
      "description": "Servidor de producción"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Autenticación basada en sesión a través de cookies."
      }
    },
    "schemas": {
...
      "Tarea": {
        "type": "object",
        "properties": {
          "pk": {
            "type": "integer",
            "description": "ID único de la tarea."
          },
          "fecha_inicio": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Fecha y hora de inicio de la tarea."
          },
          "fecha_fin": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Fecha y hora de fin de la tarea."
          },
          "descripcion": {
            "type": "string",
            "description": "Descripción detallada de la tarea."
          },
          "prioridad": {
            "type": "integer",
            "nullable": true,
            "description": "Nivel de prioridad de la tarea."
          },
          "titulo": {
            "type": "string",
            "description": "Título de la tarea."
          },
          "usuario": {
            "type": "integer",
            "description": "ID del usuario al que pertenece la tarea."
          },
          "tiempo_estimado": {
            "type": "string",
            "nullable": true,
            "description": "Tiempo estimado para completar la tarea (ej. '3 días')."
          },
          "horas": {
            "type": "integer",
            "nullable": true,
            "description": "Horas estimadas para completar la tarea."
          }
        },
        "required": [
          "pk",
          "descripcion",
          "titulo",
          "usuario"
        ]
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Mensaje de error."
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "mensaje": {
            "type": "string",
            "example": "Inicio de sesión exitoso"
          },
          "usuario": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer"
              },
              "email": {
                "type": "string"
              }
            }
          },
          "secretKeyId": {
            "type": "string",
            "description": "ID de la clave secreta de AWS S3."
          },
          "secretKey": {
            "type": "string",
            "description": "Clave secreta de AWS S3."
          }
        }
      },
      "LogoutResponse": {
        "type": "object",
        "properties": {
          "mensaje": {
            "type": "string",
            "example": "Sesión cerrada"
          }
        }
      },
...
    }
  },
  "paths": {
    "/usuarios": {
      "post": {
        "summary": "Crear un nuevo usuario",
        "tags": [
          "Autenticación"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Correo electrónico del nuevo usuario."
                  },
                  "password": {
                    "type": "string",
                    "description": "Contraseña del nuevo usuario."
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario creado exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "pk": {
                      "type": "integer"
                    },
                    "email": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Email y contraseña son requeridos.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error al hashear la contraseña o al insertar el usuario en la base de datos.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/login": {
      "post": {
        "summary": "Iniciar sesión de usuario",
        "tags": [
          "Autenticación"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Correo electrónico del usuario."
                  },
                  "password": {
                    "type": "string",
                    "description": "Contraseña del usuario."
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Inicio de sesión exitoso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "Email y contraseña requeridos.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Usuario no encontrado o contraseña incorrecta.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error al verificar contraseña o al guardar la sesión.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/logout": {
      "post": {
        "summary": "Cerrar sesión de usuario",
        "tags": [
          "Autenticación"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sesión cerrada exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponse"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado. Inicia sesión primero.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/tareas/de/{usuario}": {
      "get": {
        "summary": "Obtener todas las tareas de un usuario específico",
        "tags": [
          "Tareas"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "usuario",
            "in": "path",
            "required": true,
            "description": "ID del usuario cuyas tareas se desean obtener.",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de tareas del usuario.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Tarea"
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado. Inicia sesión primero.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "No autorizado. Solo puedes acceder a tus propias tareas.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error al obtener tareas por usuario.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/tareas/por/{id}": {
      "get": {
        "summary": "Obtener una tarea por su ID",
        "tags": [
          "Tareas"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID de la tarea a obtener.",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalles de la tarea.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tarea"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado. Inicia sesión primero.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "No autorizado. Solo puedes acceder a tus propias tareas.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Tarea no encontrada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error al obtener tarea por ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/tareas": {
      "post": {
        "summary": "Crear una nueva tarea",
        "tags": [
          "Tareas"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NuevaTarea"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tarea creada exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tarea"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado. Inicia sesión primero.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error al crear nueva tarea.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/tareas/{id}": {
      "put": {
        "summary": "Actualizar una tarea existente",
        "tags": [
          "Tareas"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID de la tarea a actualizar.",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NuevaTarea"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tarea actualizada exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mensaje": {
                      "type": "string",
                      "example": "Tarea actualizada"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado. Inicia sesión primero.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "No autorizado. Solo puedes modificar tus propias tareas.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Tarea no encontrada o no se realizaron cambios.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error al actualizar tarea.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Eliminar una tarea",
        "tags": [
          "Tareas"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID de la tarea a eliminar.",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tarea eliminada exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mensaje": {
                      "type": "string",
                      "example": "Tarea eliminada"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado. Inicia sesión primero.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "No autorizado. Solo puedes eliminar tus propias tareas.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Tarea no encontrada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error al eliminar tarea.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}